Added 'services' folder and added viewList feature
The 'services' folder contains the UserService.js file which will contain the APIs for the user.
All other APIs will be added to this folder.

The 'UserList' page has been added to the 'pages' folder. This page will display the list of users in the system.
!! It is dependent on the changed I made in the backend.

The NavBar now more closely resembles what the professor wants.

The Footer has been updated to include the correct links and the correct text.

Cleaned up the css files for pages.