Expanded on Justin's create new user page.
I tried (and failed) getting the create new user page functional as a starting point for populating the database.
It isn't the prettiest as it is just a starting point. For now it has the fields that it needs, and no errors pop up in the console. I just need to figure out why it isn't updating the database.

I also added a new service for the sign up page.