Finished up as much UI as I could
Create New User isn't working with the backend main branch merged 9pm EST. It gets an error **"loggedInUser" is null** in Spring Boot.
If we fix this it should work.

UserList is not working with the recent main branch merged 9pm EST. Not sure what is causing the issue but it worked with the main branch before the merge around 9, so if we can determine what changed then we should be able to fix it.

Admin UI is done. It isn't the prettiest but it is good enough for now.
The View All Users button works.
The Create New User button works.
The View Expired Passwrods button just brings up a modal window right now but getting it functional is low priority, as no passwords should be expired right now so we could just say that when we do the demo.

Forgot Password UI is done, logic and api are partially done.

Login still needs to be done.

NavBar is mostly done.
Right now, clicking the username or icon takes you to admin dashboard. Once we get the log in page done, we can use an api to get the username so it is dynamic. We can also get the role of the user logged in and adjust the conditional in NavBar.jsx to navigate to different dashboards depending on the user role.