Multiple updates
Login pages should all work now.
- Only issue is they all work regardless of the role, so we may just
  want to go back to having one login page since it navigates to
  different pages depending on the role of the user logging in.

NavBar is updated based on the role of the user signed in.
- We should add a Log Out button somewhere.

ForgotPassword page is now a page that verifies your credentials,
and then a page that allows you to input a new password.
- The UI is there, but there are no APIs or functions that check that
  the input is valid.

Admin Dashboard now has links in the navbar to eventually replace the cluttered dashbaord (the clutter is still there until all issues with dashboard
are resolved).
- UserList still isn't working for some reason unless "@Admin" is removed
  from the spring boot application. I'm currently looking into ways to
  resolve this issue.

CreateNewUser's API now works, but the UI needs some work.