Local Storage added; changes made to UI and API.
Still having an issue with the APIs that require user authentication:
- users/allUsers
- users/activate
- users/deactivate
I tried using localStorage to hold the user data, and it does show up in the web dev console >> Application >> Local Storage. It has the correct info required. I tried passing the user object in the request, passing the username and role as part of the request, and passing nothing as part of the request; but I had no luck.

CreateNewUser UI and API should be working now. The UI isn't great, but making it prettier is low priority.

ForgotPassword UI is fixed. Only one question is shown now, and no username input is shown.
- API for getting the security question works.
ISSUES:
- API for verifying the answer does not work (it works in postman, so it is an issue in the front-end).
- Once the verifyAnswer API is working, it should navigate the user to the EnterNewPassword component.

EnterNewPassword's UI should be fine.
The APIs still need to be set up though.