New components and forms for accounts.
Added Accounts management page for administrator.
Tabs take teh admin to the appropriate ui.
Only the add UI has been set up and I am still working on making sure only fields that make sense are there, and any fields that shouldn't be there are not.
Calendar pop up in admin account management page will pass the selecetd date to the form used in the selected tab. Currently it only does this with "View" as I wasn't sure if it was necessary for edit, add, and deactivate.

Updated the Navbar to allow log out.
Added logout functionality.

Made "/user-login" the landing page.
Got rid of admin login and manager login as the frontend will direct the user to the correct dashboard based on their credentials.