Final updates for sprint 6
Journal Attachments now store in the backend and can be viewed in the frontend through post references.
The User Login page is now formatted a little nicer, it was a bit lopsided before.
The deactivate account feature now checks that the account balance is zero, and displays and error message in a modal if the balance is not zero.
- deavtivated accounts are now displayed in a red row with strikethrough text indicating they have been deactivated.
- deactivated accounts can be reactivated from the deactivate account page.
Accountant navigation from general ledger to subledger is now working.
Accountant navigation from subledger to general ledger is now working.